Sesión 7/6/2023. @marcrodriiguez7, @jordirubio01, @atorra11. ·menu.c/h: nuevas funciones para el uso de pila (stack) implementadas para guardar los tiempos de ejecución del programa entero y de las sesiones de los usuarios. ·content.h/c: los usuarios ya pueden ver todas las publicaciones guardadas en el archivo "content.txt" (falta poder ver publicaciones de un usuario en concreto y comprobar que se pueda publicar desde cero). ·dictionary.h/c: falta comprobar que el resto de funciones se ejecuten correctamente. ·"content.txt": nuevo fichero creado para almacenar las publicaciones de los usuarios.

+++ user.h
@@ -2,6 +2,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <time.h>
+#include <math.h>
 #ifndef LAB_0_USER_H
 #define LAB_0_USER_H
 #define MAX_LENGTH 25
@@ -26,7 +27,7 @@ typedef struct{
     int type;                       // Monumento, Gastronomía, Hospedaje, Deporte, Transporte, Otros
     char location[MAX_LENGTH];      // Ubicación
     char schedule[MAX_LENGTH];      // Horarios
-    double price;                   // Precio
+    float price;                    // Precio
     int stars;                      // Valoración (0-5)
     char review[MAX_LENGTH_REVIEW]; // Reseña, Opiniones, Comentarios
     char username[MAX_LENGTH];
